[MOV] hr_recruitment_survey: split survey models into separate files
Just reordering a bit code while passing by.

Task-3093257 (Mail: The Composer Update)
