+++ addons/hr_recruitment_survey/models/__init__.py
@@ -3,4 +3,4 @@
 
 from . import hr_job
 from . import hr_applicant
-from . import survey_invite
+from . import survey_user_input
