+++ addons/hr_recruitment_survey/wizard/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from . import survey_invite
